Reorganizing so regex action is not limited to TextActions
