+++ app/src/main/java/net/gsantner/markor/format/markdown/MarkdownHighlighter.java
@@ -17,7 +17,6 @@ import android.text.InputFilter;
 import net.gsantner.markor.model.Document;
 import net.gsantner.markor.ui.hleditor.Highlighter;
 import net.gsantner.markor.ui.hleditor.HighlightingEditor;
-import net.gsantner.markor.ui.hleditor.ListHandler;
 import net.gsantner.markor.util.AppSettings;
 
 import other.writeily.format.markdown.WrMarkdownHeaderSpanCreator;
@@ -45,7 +44,11 @@ public class MarkdownHighlighter extends Highlighter {
         _highlightCodeChangeFont = _appSettings.isMarkdownHighlightCodeFontMonospaceAllowed();
         _highlightBiggerHeadings = _appSettings.isMarkdownBiggerHeadings();
         _highlightDisableCodeBlock = _appSettings.isMarkdownDisableCodeBlockHighlight();
-        setTextModifier(new ListHandler(_appSettings.isMarkdownAutoUpdateList()));
+        clearTextModifiers();
+        addTextModifier(new EmptyListRemover());
+        if (_appSettings.isMarkdownAutoUpdateList()) {
+            //addTextModifier(new OrderedListRenumberer());
+        }
     }
 
     @Override
