+++ app/src/main/java/net/gsantner/markor/ui/hleditor/HighlightingEditor.java
@@ -123,20 +123,26 @@ public class HighlightingEditor extends AppCompatEditText {
     public void enableHighlighterAutoFormat() {
         setFilters(new InputFilter[]{_hl.getAutoFormatter()});
 
-        TextWatcher modifier = (_hl != null) ? _hl.getTextModifier() : null;
-        if (modifier != null && !_appliedModifiers.contains(modifier)) {
-            addTextChangedListener(modifier);
-            _appliedModifiers.add(modifier);
+        if (_hl != null) {
+            for (TextWatcher modifier : _hl.getTextModifiers()) {
+                if (modifier != null && !_appliedModifiers.contains(modifier)) {
+                    addTextChangedListener(modifier);
+                    _appliedModifiers.add(modifier);
+                }
+            }
         }
     }
 
     public void disableHighlighterAutoFormat() {
         setFilters(new InputFilter[]{});
 
-        TextWatcher modifier = (_hl != null) ? _hl.getTextModifier() : null;
-        if (modifier != null) {
-            removeTextChangedListener(modifier);
-            _appliedModifiers.remove(modifier);
+        if (_hl != null) {
+            for (TextWatcher modifier : _hl.getTextModifiers()) {
+                if (modifier != null) {
+                    removeTextChangedListener(modifier);
+                    _appliedModifiers.remove(modifier);
+                }
+            }
         }
     }
